unzip test.zip => python3 main.py
