+++ vlad.ivanov/perc.py
@@ -3,15 +3,17 @@ from random import randint
 
 
 class Perceptron(object):
-	def __init__(self, s_len, a_len, sToA = None, aToR = None, trainA = None):
+	def __init__(self, s_len = None, a_len = None, sToA = None, aToR = None, trainA = None):
 		if sToA == None:
 			self.sToA = []
 		if aToR == None:	
 			self.aToR = []
 		if trainA == None:
 			self.trainA = []
-		self.s_len = s_len
-		self.a_len = a_len
+		if s_len == None:
+			self.s_len = 0
+		if a_len == None:
+			self.a_len = 0
 
 	def findA(self, vector):
 		temp = [0] * self.a_len
@@ -123,7 +125,9 @@ class Perceptron(object):
 		self.aToR = []
 		for t in lst[:-2]:
 			self.sToA.append(list(map(int, t[:-1].split())))
-		self.aToR = list(map(int, lst[len(lst) - 1].split()))
+		self.aToR = list(map(int, lst[-1].split()))
+		self.a_len = len(self.aToR)
+		self.s_len = len(self.sToA)
 
 	def checkTrainA(self):
 		for t in range(5000):
@@ -132,43 +136,7 @@ class Perceptron(object):
 					print(t, g)
 
 
-def createTData():
-	ans = list(map(int, open("train_labels").read().split()))
-	for i in range(NUM_TRAIN):
-		print(i)
-		x = list(map(int, open("train/" + str(i + 1) + ".txt").read().split()))
-		if ans[i] != cur_d:
-			tData.append((x, -1))
-		else:
-			tData.append((x, 1))
 
-	for i in range(NUM_TRAIN, NUM_TRAIN + 20000):
-		if ans[i] == cur_d:
-			x = list(map(int, open("train/" + str(i + 1) + ".txt").read().split()))
-			tData.append((x, 1))
-NUM_TRAIN = 20000
-cur_d = 2
-
-#tData = []
-#createTData()
-#tData = tData[::-1]
-#print(len(tData))
-#print('training')
-
-p = Perceptron(784, 3500)
-#p.train(tData)
-#p.extract('stoaandator2')
-
-p.insert('stoaandator2')
-print('yep')
-
-
-fw = open("perc_1", 'w')
-for t in range(1000):
-	if p.check(list(map(int, open("test/" + str(t + 1) + ".txt").read().split()))) == 1:
-		fw.write(str(t + 1) + ':2\n')	
-	if t % 500 == 0:
-		print(t)
